requirements.txt
asgiref==3.2.7
boto3==1.17.5
botocore==1.20.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.4
django-crispy-forms==1.9.0
django-import-export==2.4.0
django-storages==1.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
praw==6.5.1
prawcore==1.0.1
psycopg2-binary==2.8.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
s3transfer==0.3.4
sqlparse==0.3.1
stripe==2.48.0
tablib==3.0.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
whitenoise==5.0.1
xlrd==2.0.1
xlwt==1.3.0
